added changes/adjustments in line with comments
