+++ src/app/page.tsx
@@ -1,5 +1,7 @@
 import Image from "next/image";
+import ResetPasswordPage from "./reset-password/page";
 
 export default function Home() {
-  return <h1>Hello, Next.js!</h1>
+  // return <h1>Hello, Next.js!</h1>
+  return <ResetPasswordPage></ResetPasswordPage>
 }
